poprawki drobne

+++ ServerApplication/tests/auto/UnitTests/testlibs/CReceiverMock.cpp
@@ -1,45 +1,45 @@
 #include "CReceiverMock.h"
 
 QTcpSocket *CReceiverMock::GetSocket() const {
-		return 0;
+  return 0;
 }
 
 const QString CReceiverMock::GetLog() const {
-    return mLog;
+  return mLog;
 }
 
 void CReceiverMock::NewData() {
-    mLog += "->NewData()";
+  mLog += "->NewData()";
 }
 
 void CReceiverMock::RouteData(char aData) {
-		Q_UNUSED(aData);
+  Q_UNUSED(aData);
 }
 
 void CReceiverMock::ServeReceivedMessage() {
 }
 
 bool CReceiverMock::HasMessageCorrectFormat(char *aMessage) {
-		Q_UNUSED(aMessage);
-		return true;
+  Q_UNUSED(aMessage);
+  return true;
 }
 
 void CReceiverMock::ServeReceivedFileData() {
 }
 
-uint16_t CReceiverMock::CalculateFileDataChecksum(QByteArray aData) {
-		Q_UNUSED(aData);
-		return 0;
+uint32_t CReceiverMock::CalculateFileDataChecksum(QByteArray aData) {
+  Q_UNUSED(aData);
+  return 0;
 }
 
 int32_t CReceiverMock::ByteArrayToInt(QByteArray aData) {
-		Q_UNUSED(aData);
-		return 0;
+  Q_UNUSED(aData);
+  return 0;
 }
 
 void CReceiverMock::Connect(QTcpSocket *aSocket) {
-		Q_UNUSED(aSocket);
-		mLog += "->Connect(0)";  // aSocket should be is nullptr set in constructor
+  Q_UNUSED(aSocket);
+  mLog += "->Connect(0)";  // aSocket should be is nullptr set in constructor
 }
 
 void CReceiverMock::ConnectSocketSignals() {
@@ -49,9 +49,9 @@ void CReceiverMock::Disconnected() {
 }
 
 int CReceiverMock::ConvertMessageArrayToInt() {
-    return 0;
+  return 0;
 }
 
 void CReceiverMock::ResponeToClient(const char *aMessage) {
-		mLog += "->ResponeToClient(" + QString::fromStdString(aMessage) + ")";
+  mLog += "->ResponeToClient(" + QString::fromStdString(aMessage) + ")";
 }
